Refactored code - sprint 2 bringing pieces together
GUI now is done bar the lights switching, any ideas on this would be appreciated. Echo records 10 seconds of speech and passes this back as audio after finishing, we can tie this together with the search for wake word to see the feasabilty of it for hand in on thursday.